+++ R/zzz.r
@@ -1,10 +1,12 @@
-.onAttach <- function(...) {
+# this is only used during active development phases before/after CRAN releases
 
-  if (!interactive()) return()
-
-  # packageStartupMessage(paste0("cdcfluview is under *active* development. ",
-  #                              "There are *breaking changes*. ",
-  #                              "See https://github.com/hrbrmstr/cdcfluview for info/news."))
-
-}
-
+# .onAttach <- function(...) {
+#
+#   if (!interactive()) return()
+#
+#   packageStartupMessage(paste0("cdcfluview is under *active* development. ",
+#                                "There are *breaking changes*. ",
+#                                "See https://github.com/hrbrmstr/cdcfluview for info/news."))
+#
+# }
+#