+++ R/utils.r
@@ -13,3 +13,12 @@
   tbl
 
 }
+
+to_num <- function(x) {
+  x <- gsub("%", "", x, fixed=TRUE)
+  x <- gsub(">", "", x, fixed=TRUE)
+  x <- gsub("<", "", x, fixed=TRUE)
+  x <- gsub(",", "", x, fixed=TRUE)
+  x <- gsub(" ", "", x, fixed=TRUE)
+  as.numeric(x)
+}